requirements.txt
annotated-types==0.7.0
anyio==4.8.0
ardent-sdk==0.3.6
asn1crypto==1.5.1
attrs==25.1.0
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.24.0
azure-mgmt-containerservice==39.1.0
azure-mgmt-core==1.6.0
azure-storage-file-share==12.22.0
boto3==1.36.11
braintrust==0.2.6
botocore==1.36.11
cachetools==5.5.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
databricks-api==0.9.0
databricks-cli==0.18.0
databricks-sdk==0.41.0
Deprecated==1.2.18
deprecation==2.1.0
dnspython==2.7.0
docker==7.1.0
durationpy==0.10
exceptiongroup==1.2.1
execnet==2.1.1
filelock==3.18.0
fonttools==4.55.8
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
gotrue==2.12.3
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.27.0
hyperframe==6.1.0
idna==3.7
importlib_resources==6.5.2
iniconfig==2.0.0
isodate==0.7.2
jmespath==1.0.1
kiwisolver==1.4.7
kubernetes==33.1.0
matplotlib==3.9.4
motor==3.7.0
msal==1.33.0
msal-extensions==1.3.1
mysql-connector-python==9.2.0
numpy==2.0.2
oauthlib==3.3.1
packaging==24.0
pandas==2.2.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.4.0
pluggy==1.5.0
postgrest==1.1.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGithub==2.5.0
PyJWT==2.10.1
pymongo==4.11
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
pytest==8.2.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-on-whales==0.75.1
pytz==2025.1
PyYAML==6.0.2
realtime==2.6.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.11.2
seaborn==0.13.2
service-identity==24.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.17.2
sortedcontainers==2.4.0
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
tabulate==0.9.0
tomli==2.0.1
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
urllib3==1.26.20
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
